// src/api/storage-equipment/models/EquipmentItem.ts
export interface EquipmentItem {
    /**
     * 
     * @type {string}
     * @memberof EquipmentItem
     */
    id: string;
    /**
     * Name of the equipment
     * @type {string}
     * @memberof EquipmentItem
     */
    name: string;
    /**
     * Serial number of the equipment
     * @type {string}
     * @memberof EquipmentItem
     */
    serialNumber: string;
    /**
     * Manufacturer of the equipment
     * @type {string}
     * @memberof EquipmentItem
     */
    manufacturer: string;
    /**
     * Model of the equipment
     * @type {string}
     * @memberof EquipmentItem
     */
    model?: string;
    /**
     * Date when equipment was installed
     * @type {Date}
     * @memberof EquipmentItem
     */
    installationDate: Date;
    /**
     * Location of the equipment in the hospital
     * @type {string}
     * @memberof EquipmentItem
     */
    location: string;
    /**
     * Service interval in days
     * @type {number}
     * @memberof EquipmentItem
     */
    serviceInterval?: number;
    /**
     * Date of last service
     * @type {Date}
     * @memberof EquipmentItem
     */
    lastService?: Date;
    /**
     * Date of next scheduled service
     * @type {Date}
     * @memberof EquipmentItem
     */
    nextService?: Date;
    /**
     * Life expectancy in years
     * @type {number}
     * @memberof EquipmentItem
     */
    lifeExpectancy?: number;
    /**
     * 
     * @type {string}
     * @memberof EquipmentItem
     */
    status?: EquipmentItemStatusEnum;
    /**
     * Additional notes about the equipment
     * @type {string}
     * @memberof EquipmentItem
     */
    notes?: string;
}


/**
 * @export
 */
export const EquipmentItemStatusEnum = {
    Operational: 'operational',
    InRepair: 'in_repair',
    Faulty: 'faulty',
    Decommissioned: 'decommissioned'
} as const;
export type EquipmentItemStatusEnum = typeof EquipmentItemStatusEnum[keyof typeof EquipmentItemStatusEnum];